You can now clear email or location.
To clear the email address, just try set it to a blank string. That will remove
the property.

To clear the location, set all the location attributes to blank strings. (Yes I
know this is fiddly.) When setting the location you have to set all the
attributes and when clearing them you have to clear all the attributes. All or
nothing!